Moved all BT implementation to MainActivity
