+++ app/build.gradle
@@ -26,8 +26,8 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    compile 'com.akexorcist:bluetoothspp:1.0.0'
-    compile 'com.macroyau:blue2serial:0.1.5'
+    implementation 'com.akexorcist:bluetoothspp:1.0.0'
+    implementation 'com.macroyau:blue2serial:0.1.5'
 
 
 }
