+++ app/src/main/java/group2019022/me/guideme/fragments/SectionsPagerAdapter.java
@@ -1,0 +1,37 @@
+package group2019022.me.guideme.fragments;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
+ * one of the sections/tabs/pages.
+ * <p>
+ * I just moved this so MainActivity only focus on handling the Bluetooth connection.
+ */
+public class SectionsPagerAdapter extends FragmentPagerAdapter {
+
+    private final List<Fragment> fragments = new ArrayList<>();
+
+    public SectionsPagerAdapter(FragmentManager fm) {
+        super(fm);
+        fragments.add(new Tab1Dashboard());
+        fragments.add(new Tab2Settings());
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        return fragments.get(position);
+    }
+
+    @Override
+    public int getCount() {
+        // Show 3 total pages.
+        return fragments.size();
+    }
+}
+
