+++ app/src/main/java/group2019022/me/guideme/fragments/Tab1Dashboard.java
@@ -1,14 +1,13 @@
-package group2019022.me.guideme;
+package group2019022.me.guideme.fragments;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.SeekBar;
-import android.widget.TextView;
 
-public class Tab1Dashboard extends Fragment{
+import group2019022.me.guideme.R;
+
+public class Tab1Dashboard extends BaseFragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
