+++ app/src/main/res/layout/activity_main.xml
@@ -31,7 +31,7 @@
                 android:layout_gravity="center"
                 android:text="@string/app_name"
                 android:textColor="@color/textColorBlack"
-                android:textSize="20sp"/>
+                android:textSize="20sp" />
 
 
         </android.support.v7.widget.Toolbar>
@@ -48,14 +48,14 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="@string/tab_text_1"
-                android:textSize="19sp"/>
+                android:textSize="19sp" />
 
             <android.support.design.widget.TabItem
                 android:id="@+id/tabItem2"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="@string/tab_text_2"
-                android:textSize="19sp"/>
+                android:textSize="19sp" />
 
         </android.support.design.widget.TabLayout>
     </android.support.design.widget.AppBarLayout>
@@ -65,5 +65,18 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior" />
+
+    <!-- This could be prettier of course -->
+    <Button
+        android:id="@+id/button"
+        android:layout_width="305dp"
+        android:layout_height="60dp"
+        android:layout_gravity="bottom|center_horizontal"
+        android:layout_marginStart="48dp"
+        android:layout_marginTop="56dp"
+        android:layout_marginEnd="48dp"
+        android:layout_marginBottom="@dimen/activity_vertical_margin"
+        android:text="@string/find_device"
+        android:textSize="14sp" />
 
 </android.support.design.widget.CoordinatorLayout>
