+++ app/src/main/res/layout/tab_1_dashboard.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Tab1Dashboard">
+    tools:context=".fragments.Tab1Dashboard">
 
     <TextView
         android:id="@+id/textView10"
