+++ app/src/main/res/layout/tab_2_settings.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Tab2Settings">
+    tools:context=".fragments.Tab2Settings">
 
     <TextView
         android:id="@+id/textView"
@@ -25,8 +25,8 @@
         android:id="@+id/textView2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginEnd="47dp"
         android:layout_marginStart="72dp"
+        android:layout_marginEnd="47dp"
         android:text="@string/vibration_default"
         android:textColor="@color/colorAccent"
         android:textSize="19sp"
@@ -39,9 +39,9 @@
         android:id="@+id/seekBar"
         android:layout_width="0dp"
         android:layout_height="44dp"
-        android:layout_marginEnd="36dp"
         android:layout_marginStart="36dp"
         android:layout_marginTop="35dp"
+        android:layout_marginEnd="36dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView" />
@@ -61,8 +61,8 @@
         android:id="@+id/textView4"
         android:layout_width="0dp"
         android:layout_height="18dp"
-        android:layout_marginEnd="47dp"
         android:layout_marginTop="16sp"
+        android:layout_marginEnd="47dp"
         android:text="@string/vibration_max"
         android:textColor="@color/textColorBlack"
         app:layout_constraintEnd_toEndOf="parent"
@@ -85,8 +85,8 @@
         android:id="@+id/textView6"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginEnd="47dp"
         android:layout_marginStart="72dp"
+        android:layout_marginEnd="47dp"
         android:text="@string/distance_default"
         android:textColor="@color/colorAccent"
         android:textSize="19sp"
@@ -98,9 +98,9 @@
         android:id="@+id/seekBar2"
         android:layout_width="0dp"
         android:layout_height="43dp"
-        android:layout_marginEnd="36dp"
         android:layout_marginStart="36dp"
         android:layout_marginTop="35dp"
+        android:layout_marginEnd="36dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView5" />
@@ -109,8 +109,8 @@
         android:id="@+id/textView7"
         android:layout_width="0dp"
         android:layout_height="18dp"
-        android:layout_marginEnd="47dp"
         android:layout_marginTop="16dp"
+        android:layout_marginEnd="47dp"
         android:text="@string/distance_max"
         android:textColor="@color/textColorBlack"
         app:layout_constraintEnd_toEndOf="parent"
@@ -125,18 +125,5 @@
         android:text="@string/distance_min"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/seekBar2" />
-
-    <Button
-        android:id="@+id/button"
-        android:layout_width="305dp"
-        android:layout_height="60dp"
-        android:layout_marginEnd="48dp"
-        android:layout_marginStart="48dp"
-        android:layout_marginTop="56dp"
-        android:text="@string/find_device"
-        android:textSize="14sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView8" />
 
 </android.support.constraint.ConstraintLayout>
